fetchgit: use posix_spawn (vfork really)
Plain fork() can fail with ENOMEM and it does in all non-trivial
deployments. vfork() is removed in POSIX.1-2008, so posix_spawn()
is the way to go. As GLibc implements posix_spawn() with either
fork() or vfork(), we make sure that vfork() is used.

Note that Nix itself uses vfork() on Linux.